+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Add :meth:`pymaf.MafFrame.matrix_prevalence` method.
 * Add :meth:`pymaf.MafFrame.plot_regplot` method.
 * Add ``samples`` argument to :meth:`pymaf.MafFrame.plot_snvclss` method.
+* Add :meth:`pymaf.MafFrame.plot_evolution` method.
 
 0.16.0 (2021-07-02)
 -------------------